// src/lib/interfaces/NewArticuloStock.ts
// Generated by https://quicktype.io

import { TipoImpuesto, UnidadMedida } from '@core/interfaces';

export interface INewArticuloStock {
  descripcion: string;
  precioVenta: number;
  unidadMedida: UnidadMedida;
  lineaArticulo: string;
  tipoImpuesto: TipoImpuesto;
  codigoBarra: string;
  codigo: number;
  estado: boolean;
}